+++ curves/src/templates/twisted_edwards_extended/mod.rs
@@ -149,7 +149,7 @@ impl<P: Parameters> AffineCurve for GroupAffine<P> {
         copy
     }
 
-    fn mul<S: Into<<Self::ScalarField as PrimeField>::BigInt>>(&self, by: S) -> GroupProjective<P> {
+    fn mul<S: Into<<Self::ScalarField as PrimeField>::BigInteger>>(&self, by: S) -> GroupProjective<P> {
         self.mul_bits(BitIterator::new(by.into()))
     }
 
@@ -525,7 +525,7 @@ impl<P: Parameters> ProjectiveCurve for GroupProjective<P> {
         self.z = f * &g;
     }
 
-    fn mul_assign<S: Into<<Self::ScalarField as PrimeField>::BigInt>>(&mut self, other: S) {
+    fn mul_assign<S: Into<<Self::ScalarField as PrimeField>::BigInteger>>(&mut self, other: S) {
         let mut res = Self::zero();
 
         let mut found_one = false;
@@ -549,7 +549,7 @@ impl<P: Parameters> ProjectiveCurve for GroupProjective<P> {
         (*self).into()
     }
 
-    fn recommended_wnaf_for_scalar(scalar: <Self::ScalarField as PrimeField>::BigInt) -> usize {
+    fn recommended_wnaf_for_scalar(scalar: <Self::ScalarField as PrimeField>::BigInteger) -> usize {
         P::empirical_recommended_wnaf_for_scalar(scalar)
     }
